Move code from lib.rs to frame.rs
